agent/requirements.txt
mcp
langchain
langgraph
convex
chromadb
google-generativeai
python-dotenv
pydantic==2.10.6
pydantic-settings
